only ship the fastest parser (handwritten with built-in re) and only include always correct tatsu for testing
