+++ tests/valuetype/text.py
@@ -2,12 +2,17 @@ import attr
 import pytest
 from hypothesis import given
 
-from ics.grammar import ContentLine, string_to_container
+from ics.contentline import ContentLine, string_to_containers, Parser
 from ics.valuetype.text import TextConverter
 # Text may be comma-separated multi-value but is never quoted, with the characters [\\;,\n] escaped
-from tests.grammar import VALUE
+from tests.contentline import VALUE
 
 TextConv: TextConverter = TextConverter.INST
+
+
+def parse_contentline(line: str) -> ContentLine:
+    cl, = Parser.lines_to_contentlines(Parser.string_to_lines(line))
+    return cl
 
 
 @pytest.mark.parametrize("inp_esc, out_uesc", [
@@ -17,7 +22,8 @@ TextConv: TextConverter = TextConverter.INST
     ),
     (
             "DESCRIPTION;ALTREP=\"cid:part1.0001@example.org\":The Fall'98 Wild Wizards Conference - - Las Vegas\\, NV\\, USA",
-            ContentLine("DESCRIPTION", {"ALTREP": ["cid:part1.0001@example.org"]}, value="The Fall'98 Wild Wizards Conference - - Las Vegas, NV, USA")
+            ContentLine("DESCRIPTION", {"ALTREP": ["cid:part1.0001@example.org"]},
+                        value="The Fall'98 Wild Wizards Conference - - Las Vegas, NV, USA")
     ),
     (
             "TEST:abc\\r\\n\\,\\;:\"\t=xyz",
@@ -25,13 +31,13 @@ TextConv: TextConverter = TextConverter.INST
     ),
 ])
 def test_example_text_recode(inp_esc, out_uesc):
-    par_esc = ContentLine.parse(inp_esc)
+    par_esc = parse_contentline(inp_esc)
     par_uesc = attr.evolve(par_esc, value=TextConv.parse(par_esc.value))
     out_esc = attr.evolve(out_uesc, value=TextConv.serialize(out_uesc.value))
     assert par_uesc == out_uesc
     ser_esc = out_esc.serialize()
     assert inp_esc == ser_esc
-    assert string_to_container(inp_esc) == [par_esc]
+    assert list(string_to_containers(inp_esc)) == [par_esc]
 
 
 # TODO list examples ("RESOURCES:EASEL,PROJECTOR,VCR", ContentLine("RESOURCES", value="EASEL,PROJECTOR,VCR"))
@@ -53,13 +59,14 @@ def test_broken_escape():
         TextConv.unescape_text("abc;def")
     assert e.match("unescaped character")
 
+
 def test_trailing_escape_value_list():
-    cl1 = ContentLine.parse("TEST:this is,a list \\, with a\\\\,trailing escape\\")
+    cl1 = parse_contentline("TEST:this is,a list \\, with a\\\\,trailing escape\\")
     with pytest.raises(ValueError) as excinfo:
         list(TextConv.split_value_list(cl1.value))
     assert "not end with an escape sequence" in str(excinfo.value)
 
-    cl2 = ContentLine.parse("TEST:this is,a list \\, with a\\\\,trailing escape")
+    cl2 = parse_contentline("TEST:this is,a list \\, with a\\\\,trailing escape")
     assert list(TextConv.split_value_list(cl2.value)) == \
            ["this is", "a list \\, with a\\\\", "trailing escape"]
     assert [TextConv.parse(v) for v in TextConv.split_value_list(cl2.value)] == \
@@ -71,10 +78,10 @@ def test_any_text_value_recode(value):
     esc = TextConv.serialize(value)
     assert TextConv.parse(esc) == value
     cl = ContentLine("TEST", value=esc)
-    assert ContentLine.parse(cl.serialize()) == cl
-    assert string_to_container(cl.serialize()) == [cl]
+    assert parse_contentline(cl.serialize()) == cl
+    assert list(string_to_containers(cl.serialize())) == [cl]
     vals = [esc, esc, "test", esc]
     cl2 = ContentLine("TEST", value=TextConv.join_value_list(vals))
     assert list(TextConv.split_value_list(cl2.value)) == vals
-    assert ContentLine.parse(cl.serialize()) == cl
-    assert string_to_container(cl.serialize()) == [cl]
+    assert parse_contentline(cl.serialize()) == cl
+    assert list(string_to_containers(cl.serialize())) == [cl]
